add support for win32, split generated WINAPI impl's into architectures

+++ unistd32_windows.go
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
-// +build amd64 amd64p32 arm64 mips64 mips64le mips64p32 mips64p32le ppc64 sparc64s
+// +build 386 arm arm64be armbe mips mipsle ppc ppc64le s390 s390x sparc
 
 package crt
 
@@ -15,7 +15,7 @@ import (
 )
 
 // ssize_t read(int fd, void *buf, size_t count);
-func Xread(tls *TLS, fd int32, buf unsafe.Pointer, count uint64) int64 { //TODO stdin
+func Xread(tls *TLS, fd int32, buf unsafe.Pointer, count uint64) int32 { //TODO stdin
 	slice := (*[math.MaxInt32]byte)(unsafe.Pointer(buf))[:count]
 	r, err := syscall.Read(syscall.Handle(uintptr(fd)), slice)
 	if strace {
@@ -24,5 +24,5 @@ func Xread(tls *TLS, fd int32, buf unsafe.Pointer, count uint64) int64 { //TODO 
 	if err != nil {
 		tls.setErrno(err)
 	}
-	return int64(r)
+	return int32(r)
 }
